batch/requirements.txt
celery==5.3.4
docx2txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
Flask-RESTful==0.3.9
flask-socketio
gunicorn==21.2.0
langchain==0.3.19
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-core==0.3.39
langchain-groq==0.2.4
langchain_google_genai
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.10
markdown
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0
psycopg2-binary==2.9.9
pypdf
python-dotenv==1.0.0
redis==5.0.1
requests
sqlalchemy
unstructured
chromadb